fix: refactor sheets grid range resolution, tighten validation, add tests
Extract resolveGridRange helper to deduplicate sheet lookup logic across
addDataValidation, protectRange, and clearProtectedRange. Make values
required for addDataValidation. Add int constraint to reorderSlides
insertionIndex. Add validation error tests for deleteSheet,
addDataValidation, and deleteGoogleSlide.

+++ src/tools/slides.ts
@@ -137,7 +137,7 @@ const DuplicateSlideSchema = z.object({
 const ReorderSlidesSchema = z.object({
   presentationId: z.string().min(1, "Presentation ID is required"),
   slideObjectIds: z.array(z.string().min(1)).min(1, "At least one slide object ID is required"),
-  insertionIndex: z.number().min(0)
+  insertionIndex: z.number().int().min(0)
 });
 
 const ReplaceAllTextInSlidesSchema = z.object({
